refactor(PulsingDot): each is an island
Each dot hydrates itself independently,
relying on the React API everywhere
we manipulate the DOM.
