+++ dotcom-rendering/src/components/ArticlePage.tsx
@@ -9,7 +9,6 @@ import type { NavType } from '../model/extract-nav';
 import type { FEArticleType } from '../types/frontend';
 import type { RenderingTarget } from '../types/renderingTarget';
 import { AlreadyVisited } from './AlreadyVisited.importable';
-import { AnimatePulsingDots } from './AnimatePulsingDots.importable';
 import { BrazeMessaging } from './BrazeMessaging.importable';
 import { FocusStyles } from './FocusStyles.importable';
 import { Island } from './Island';
@@ -92,11 +91,6 @@ export const ArticlePage = (props: WebProps | AppProps) => {
 			{(format.design === ArticleDesign.LiveBlog ||
 				format.design === ArticleDesign.DeadBlog) && (
 				<SkipTo id={'key-events-carousel'} label="Skip to key events" />
-			)}
-			{format.design === ArticleDesign.LiveBlog && (
-				<Island clientOnly={true} deferUntil="idle">
-					<AnimatePulsingDots />
-				</Island>
 			)}
 			{renderingTarget === 'Web' && (
 				<>
